Started restructure of directory structure in daemon
