+++ muondetector-daemon/include/custom_io_operators.h
