+++ muondetector-daemon/include/i2c/addresses.h
