+++ muondetector-daemon/include/i2c/i2cdevices/hmc5883/hmc5883.h
