+++ muondetector-daemon/include/i2c/i2cdevices/i2cdevice.h
