+++ muondetector-shared/CMakeLists.txt
@@ -38,6 +38,7 @@ add_compile_options(
 set(PROJECT_SOURCE_FILES
     "${PROJECT_SRC_DIR}/tcpconnection.cpp"
     "${PROJECT_SRC_DIR}/tcpmessage.cpp"
+    "${PROJECT_SRC_DIR}/mqtthandler.cpp"
     )
 
 set(PROJECT_HEADER_FILES
@@ -46,6 +47,7 @@ set(PROJECT_HEADER_FILES
     "${PROJECT_HEADER_DIR}/histogram.h"
     "${PROJECT_HEADER_DIR}/tcpconnection.h"
     "${PROJECT_HEADER_DIR}/tcpmessage.h"
+    "${PROJECT_HEADER_DIR}/mqtthandler.h"
     "${PROJECT_HEADER_DIR}/tcpmessage_keys.h"
     "${PROJECT_HEADER_DIR}/ublox_messages.h"
     "${PROJECT_HEADER_DIR}/ubx_msg_key_name_map.h"
@@ -57,12 +59,18 @@ add_definitions(-DMUONDETECTOR_SHARED_LIBRARY)
 
 add_library(${PROJECT_NAME} SHARED ${PROJECT_SOURCE_FILES} ${PROJECT_HEADER_FILES})
 target_include_directories(${PROJECT_NAME} PUBLIC ${PROJECT_HEADER_DIR})
-target_link_libraries(${PROJECT_NAME} Qt5::Network)
+target_link_libraries(${PROJECT_NAME} Qt5::Network paho-mqtt3c paho-mqtt3a paho-mqtt3cs paho-mqtt3as paho-mqttpp3)
+target_include_directories(${PROJECT_NAME} PUBLIC
+    $<BUILD_INTERFACE:/usr/local/include/mqtt>
+    $<INSTALL_INTERFACE:include/mqtt>)
 
 add_library(muondetector_static STATIC ${PROJECT_SOURCE_FILES} ${PROJECT_HEADER_FILES})
 set_target_properties(muondetector_static PROPERTIES OUTPUT_NAME ${PROJECT_NAME})
 target_include_directories(muondetector_static PUBLIC ${PROJECT_HEADER_DIR})
-target_link_libraries(muondetector_static Qt5::Network)
+target_link_libraries(muondetector_static Qt5::Network paho-mqtt3c paho-mqtt3a paho-mqtt3cs paho-mqtt3as paho-mqttpp3)
+target_include_directories(muondetector_static PUBLIC
+    $<BUILD_INTERFACE:/usr/local/include/mqtt>
+    $<INSTALL_INTERFACE:include/mqtt>)
 
 install(TARGETS ${PROJECT_NAME} muondetector_static DESTINATION lib)
 install(FILES ${PROJECT_HEADER_FILES} DESTINATION include)
